CD_FINNGEN {topr}	R Documentation

Finngen r7 Crohn‘s disease (K11_CROHNS)

Description

Dataset retrieved from the Finngen database (version 7) including 3147 crohn´s cases (K50) and 296,100 controls. The dataset has been filtered on variants with P <1e-03. FinnGen data are publicly available and were downloaded from https://finngen.fi.

Usage

CD_FINNGEN

Format

A data frame with 32,303 rows and 8 variables:

CHROM

Chromosome, written as for example chr1 or 1

POS

genetic position of the variant

REF

the reference allele

ALT

the alternative allele

P

P-value from Plink run, additive model, regression model GLM_FIRTH

BETA

Variant effect

ID

Variant identifier, e.g. rsid

AF

Allele frequency

Source

Crohn's K50 (K11_CROHNS), only including variants with P<1e-03

[Package topr version 2.0.2 Index]