R2_CD_UKBB {topr}	R Documentation

Example dataset including the R2 column for the locuszoom plot function

Description

The dataset is a subset of CD_UKBB and only includes variants above and near the IL23R gene on chromosome 1

Usage

R2_CD_UKBB

Format

A data frame with 329 rows and 5 variables:

CHROM

Chromosome, written as for example chr1 or 1

POS

genetic position of the variant

ID

Variant identifier, e.g. rsid

P

P-value from Plink run, additive model, regression model GLM_FIRTH

R2

variant correlation (r^2)

Source

A subset of the CD_UKBB dataset

[Package topr version 2.0.2 Index]