UC_UKBB {topr}	R Documentation

UKBB Ulcerative colitis (ICD 10 code K51)

Description

Dataset retrieved from the UK biobank including of 5,452 UC cases (K51) and 481,862 controls. The dataset has been filtered on variants with P<1e-03.

Usage

UC_UKBB

Format

A data frame with 45,012 rows and 8 variables

CHROM

Chromosome, written as for example chr1 or 1

POS

genetic position of the variant

REF

the reference allele

ALT

the alternative allele

ID

Variant identifier, e.g. rsid

P

P-value from Plink run, additive model, regression model GLM_FIRTH

OR

Odds Ratio

AF

Allele frequency

Source

Ulcerative Colitis UKBB ICD10 code K51, only including variants with P<1e-03

[Package topr version 2.0.2 Index]