annotate_with_nearest_gene {topr}	R Documentation

Get the nearest gene for one or more snps

Description

annotate_with_nearest_gene() Annotate the variant/snp with their nearest gene Required parameters is a dataframe of SNPs (with the columns CHROM and POS)

Usage

annotate_with_nearest_gene(
  variants,
  protein_coding_only = FALSE,
  build = 38,
  .chr_map = NULL
)

Arguments

variants
a dataframe of variant positions (CHROM and POS)

protein_coding_only
Logical, if set to TRUE only annotate with protein coding genes (the default value is FALSE)

build
A number representing the genome build. Set to 37 to change to build (GRCh37). The default is build 38 (GRCh38).

.chr_map
An internally used list which maps chromosome names to numbers.

Value

the input dataframe with Gene_Symbol as an additional column

Examples

## Not run: 
variants <-get_lead_snps(CD_UKBB)
annotate_with_nearest_gene(variants)

## End(Not run)

[Package topr version 2.0.2 Index]