Chromosome plot

chromplot() displays the association results for an entire chromosome. Required parameter is a dataset (dataframe) containing the association data (with columns CHROM,POS,P in upper or lowercase)

All other input parameters are optional. A dataset including results from more than one chromosome can be used as input if provided with the chr parameter.

E.g. chromplot(multi_chr_dataset,chr = "chr15")

chromplot(
  dat,
  annotation_thresh = NULL,
  title = "",
  size = 1.2,
  shape = 19,
  alpha = 1,
  color = c("darkblue", "#E69F00", "#00AFBB", "#999999", "#FC4E07", "darkorange1"),
  label_size = 3.5,
  sign_thresh = 5.1e-09,
  sign_thresh_color = "red",
  legend_position = "right",
  .checked = FALSE,
  show_xaxis = TRUE,
  axis_text_size = 11,
  title_text_size = 12,
  axis_title_size = 13,
  legend_title_size = 12,
  legend_text_size = 12,
  variant_ids_color = "red",
  legend_labels = NULL,
  legend_name = "Data",
  xmin = 0,
  highlight_genes = NULL,
  highlight_genes_ypos = NULL,
  highlight_genes_color = NULL,
  variants = NULL,
  xmax = NULL,
  ymin = NULL,
  ymax = NULL,
  chr = NULL,
  vline = NULL,
  variant_ids = NULL
)

Arguments

annotation_thresh
Display annotation for variants with p-values below this threshold

title
Plot title (optional)

size
Optional parameter setting the size of the plot points (default: size=1.2)

shape
A number of vector of numers setting the shape of the plotted points

alpha
A number or vector of numbers setting the transparancy of the plotted points

color
Optional parameter setting the color of the plot points (default: color="darkblue")

label_size
Optional parameter to set the size of the plot labels (default: label_size=3.5)

sign_thresh
Optional parameter setting the threshold of the dashed red horizontal line representing the significance threshold (default: sign_thresh=5.1e-9). Multiple thresholds can be provided in a vector, e.g sign_thresh=c(5.1e-9,1.0e-6)). Set this parameter to NULL if you dont want this line to appear at all sign_thresh=NULL

sign_thresh_color
set the color of the significance threshold line or lines

legend_position
Top,bottom,left or right

show_xaxis
show the xaxis

axis_text_size
Text size of the x and y axes tick labels (default: 12)

title_text_size
Text size of the plot title (default: 13)

axis_title_size
Text size of the x and y title labels (default: 12)

legend_title_size
Text size of the legend title

legend_text_size
Text size of the legend text

variant_ids_color
The color of the variants in variants_id (default color is red)

legend_labels
Legend labels

legend_name
Change the name of the legend (default: None)

xmin, xmax
Parameters setting the chromosomal range to display on the x-axis

highlight_genes
A vector of genes or genes to highlight the datapoints for on the plot

highlight_genes_ypos
Display the genes at this position on the y-axis (default value is 1)

highlight_genes_color
Colors for the hihglighted genes (default: green)

variants
Optional parameter, a dataset (required columns are CHROM,POS,P) with a list of variants to display and label on the plot. Variants will be labeled with Gene_Symbol if present in the dataframe, otherwise with ID. If neither Gene_Symbol nor ID are in the dataset, the variant will be labelled with its position (CHROM and POS)

ymin, ymax
Optional parameters, min and max of the y-axis, (default values: ymin=0, ymax=max(-log10(df$P)))

chr
The chromosome to plot (i.e. chr15), only required if the input dataframe contains results from more than one chromosome

vline
Parameter (optional) to add a vertical line to the plot at a specific chromosomal position, e.g vline=204000066. Multiple values can be provided in a vector, e.g vline=c(204000066,100500188)

variant_ids
A vector of variant ids to highlight on the plot, e.g. variant_ids=c("rs1234, rs45898")

df
Dataframe (required columns are CHROM,POS,P), in upper- or lowercase) to use for the plot. See examples on how to retrieve the dataset with GOR.

rect
Rectangle to add to the plot

rect_range
Sets the size of the rectangle, default is 500 kb

variants_color
One color or a vector of colors in case multiple variant dataframes are provided

variants_size
A number or a vector of numbers (default: 1.2)

variants_alpha
A number or a vector of numbers (default: 1)

variants_shape
A number or a vector of numbers (default: 19)

geneplot_label_size
Size of the labels for the genes on the gene and exon plots

Value

a chromosome plot (ggplot object)

Examples

if (FALSE) {
data(gwas_CD)
snps_CD <- get_best_snp_per_MB(gwas_CD, thresh = 1e-09, region = 10000000)
CHR="chr16"
chromplot(gwas_CD,chr=CHR, variants = snps_CD, annotation_thresh = 5e-09)
}