get_gene {topr}	R Documentation

Get the genetic position of a gene by gene name

Description

get_gene() Get the gene coordinates for a gene Required parameter is gene name

This method is deprecated and will be removed in future versions. use get_gene_coords instead.

Usage

get_gene(gene_name, chr = NULL, build = 38)

Arguments

gene_name
A string representing a gene name (e.g. "FTO")

chr
A string, search for the genes on this chromosome only, (e.g chr="chr1")

build
A string, genome build, choose between builds 37 (GRCh37) and 38 (GRCh38) (default is 38)

Value

Dataframe with the gene name and its genetic coordinates

Examples

## Not run: 
get_gene("FTO")

## End(Not run)

[Package topr version 2.0.2 Index]