get_genes_by_Gene_Symbol {topr}	R Documentation

Get the genetic position of a gene by its gene name

Description

get_genes_by_Gene_Symbol() Get genes by their gene symbol/name Required parameters is on gene name or a vector of gene names

Usage

get_genes_by_Gene_Symbol(genes, chr = NULL, build = 38)

Arguments

genes
A string or vector of strings representing gene names, (e.g. "FTO") or (c("FTO","NOD2"))

chr
A string, search for the genes on this chromosome only, (e.g chr="chr1")

build
A string, genome build, choose between builds 37 (GRCh37) and 38 (GRCh38) (default is 38)

Value

Dataframe of genes

Examples

## Not run: 
  get_genes_by_Gene_Symbol(c("FTO","THADA"))

## End(Not run)

[Package topr version 2.0.2 Index]