get_genes_in_region {topr}	R Documentation

Get SNPs/variants within region

Description

get_genes_in_region()

Usage

get_genes_in_region(
  chr = chr,
  xmin = xmin,
  xmax = xmax,
  protein_coding_only = F,
  show_exons = F,
  show_genes = T,
  build = 38,
  region = NULL
)

Arguments

chr
A string, chromosome (e.g. chr16)

xmin
An integer representing genetic position

xmax
An integer representing genetic position

protein_coding_only
A logical scalar, if TRUE, only protein coding genes are used for annotation

show_exons
Deprecated : A logical scalar, show exons instead of genes (default show_exons=FALSE)

show_genes
A logical scalar, show genes instead of exons (default show_genes=FALSE)

build
A number representing the genome build or a data frame. Set to 37 to change to build (GRCh37). The default is build 38 (GRCh38).

region
A string representing the genetic region (e.g chr16:50693587-50734041)

Value

the genes the requested region

Examples

## Not run: 
get_genes_in_region(region="chr16:50593587-50834041")

## End(Not run)

[Package topr version 2.0.2 Index]