get_overlapping_snps_by_pos {topr}	R Documentation

Get variants that overlap between two datasets

Description

get_overlapping_snps_by_pos()

This method is deprecated and will be removed in future versions. use match_by_pos instead.

Usage

get_overlapping_snps_by_pos(df1, df2, verbose = F)

Arguments

df1
A dataframe of variants, has to contain CHROM and POS

df2
A dataframe of variants, has to contain CHROM and POS

verbose
A logical scalar (default: FALSE). Assign to TRUE to get information on which alleles are matched and which are not.

Value

The input dataframe containing only those variants with matched alleles in the snpset

Examples

## Not run: 
get_overlapping_snps_by_pos(dat1, dat2)

## End(Not run)

[Package topr version 2.0.2 Index]