get_sign_and_sugg_loci {topr}	R Documentation

Get the index/lead variants

Description

get_lead_snps() Get the top variants within 1 MB windows of the genome with association p-values below the given threshold

Usage

get_sign_and_sugg_loci(
  df,
  genome_wide_thresh = 5e-08,
  suggestive_thresh = 1e-06,
  flank_size = 1e+06,
  region_size = 1e+06
)

Arguments

df
Dataframe, GWAS summary statistics

genome_wide_thresh
A number. P-value threshold for genome wide significant loci (5e-08 by default)

suggestive_thresh
A number. P-value threshold for suggestive loci (1e-06 by default)

flank_size
A number (default = 1e6). The size of the flanking region for the significant and suggestitve snps.

region_size
A number (default = 1e6). The size of the region for top snp search. Only one snp per region is returned.

Value

List of genome wide and suggestive loci.

Examples

## Not run: 
get_sign_and_sugg_loci(CD_UKBB)

## End(Not run)

[Package topr version 2.0.2 Index]