get_snpset {topr}	R Documentation

Create a dataframe that can be used as input for making effect plots

Description

get_snpset()

Usage

get_snpset(
  df1,
  df2,
  thresh = 5e-08,
  protein_coding_only = TRUE,
  region_size = 1e+06,
  verbose = NULL,
  show_full_output = FALSE,
  build = 38,
  format = "wide"
)

Arguments

df1
The dataframe to extract the top snps from (with p-value below thresh)

df2
The dataframe in which to search for overlapping SNPs from dataframe1

thresh
A number. P-value threshold, only extract variants with p-values below this threshold (5e-08 by default)

protein_coding_only
Logical, set this variable to TRUE to only use protein_coding genes for annotation

region_size
An integer (default = 20000000) (or a string represented as 200kb or 2MB) indicating the window size for variant labeling. Increase this number for sparser annotation and decrease for denser annotation.

verbose
Logical, (default: FALSE). Assign to TRUE to get information on which alleles are matched and which are not.

show_full_output
A logical scalar (default:FALSE). Assign to TRUE to show the full output from this function

build
A string, genome build, choose between builds 37 (GRCh37) and 38 (GRCh38) (default is 38)

format
A string, representing either wide or long format (default : "wide"). By default a snpset created from two dataframes is returned in a wide format.

Value

Dataframe of overlapping snps (snpset)

Examples

## Not run: 
CD_UKBB_index_snps <-get_lead_snps(CD_UKBB)
get_snpset(CD_UKBB_index_snps, CD_FINNGEN)

## End(Not run)

[Package topr version 2.0.2 Index]