get_top_snp {topr}	R Documentation

Get the top hit from the dataframe

Description

get_top_snp() Get the top hit from the dataframe All other input parameters are optional

Usage

get_top_snp(df, chr = NULL)

Arguments

df
Dataframe containing association results

chr
String, get the top hit in the data frame for this chromosome. If chromosome is not provided, the top hit from the entire dataset is returned.

Value

Dataframe containing the top hit

Examples

## Not run: 
get_top_snp(CD_UKBB, chr="chr1")

## End(Not run)

[Package topr version 2.0.2 Index]