Manhattan plot

manhattan() displays the association results for the entire genome Required parameter is at least one dataset (dataframe) containing the association data (with columns CHROM,POS,P in upper or lowercase)

All other input parameters are optional

manhattan(
  df,
  ntop = 3,
  title = "",
  color = c("darkblue", "#E69F00", "#00AFBB", "#999999", "#FC4E07", "darkorange1"),
  sign_thresh = 5e-09,
  label_size = 3,
  size = 1,
  shape = 19,
  alpha = 1,
  highlight_genes_color = "green",
  variants_size = 1.2,
  variants_alpha = 1,
  variants_shape = 19,
  axis_text_size = 11,
  axis_title_size = 12,
  title_text_size = 13,
  legend_title_size = 12,
  legend_text_size = 12,
  rect = NULL,
  protein_coding_only = FALSE,
  variants = NULL,
  legend_labels = NULL,
  legend_name = NULL,
  annotation_thresh = NULL,
  ymin = NULL,
  ymax = NULL,
  variants_color = NULL,
  highlight_genes = NULL,
  highlight_genes_ypos = NULL,
  legend_position = NULL
)

Arguments

df
Dataframe (required columns are CHROM,POS,P), in upper- or lowercase) to use for the plot. See examples on how to retrieve the dataset with GOR.

ntop
Number of datasets (GWASes) to show on the top plot

title
Plot title (optional)

color
Optional parameter setting the color of the plot points (default: color="darkblue")

sign_thresh
Optional parameter setting the threshold of the dashed red horizontal line representing the significance threshold (default: sign_thresh=5.1e-9). Multiple thresholds can be provided in a vector, e.g sign_thresh=c(5.1e-9,1.0e-6)). Set this parameter to NULL if you dont want this line to appear at all sign_thresh=NULL

label_size
Optional parameter to set the size of the plot labels (default: label_size=3.5)

size
Optional parameter setting the size of the plot points (default: size=1.2)

shape
A number of vector of numers setting the shape of the plotted points

alpha
A number or vector of numbers setting the transparancy of the plotted points

highlight_genes_color
Colors for the hihglighted genes (default: green)

variants_size
A number or a vector of numbers (default: 1.2)

variants_alpha
A number or a vector of numbers (default: 1)

variants_shape
A number or a vector of numbers (default: 19)

axis_text_size
Text size of the x and y axes tick labels (default: 12)

axis_title_size
Text size of the x and y title labels (default: 12)

title_text_size
Text size of the plot title (default: 13)

legend_title_size
Text size of the legend title

legend_text_size
Text size of the legend text

rect
Rectangle to add to the plot

protein_coding_only
Set this parameter to TRUE to only use protein coding genes for annotation

variants
Optional parameter, a dataset (required columns are CHROM,POS,P) with a list of variants to display and label on the plot. Variants will be labeled with Gene_Symbol if present in the dataframe, otherwise with ID. If neither Gene_Symbol nor ID are in the dataset, the variant will be labelled with its position (CHROM and POS)

legend_labels
Legend labels

legend_name
Change the name of the legend (default: None)

annotation_thresh
Display annotation for variants with p-values below this threshold

ymin
Optional parameters, min and max of the y-axis, (default values: ymin=0, ymax=max(-log10(df$P)))

ymax
Optional parameters, min and max of the y-axis, (default values: ymin=0, ymax=max(-log10(df$P)))

variants_color
One color or a vector of colors in case multiple variant dataframes are provided

highlight_genes
A vector of genes or genes to highlight the datapoints for on the plot

highlight_genes_ypos
Display the genes at this position on the y-axis (default value is 1)

legend_position
Top,bottom,left or right

Value

plots using egg (https://cran.r-project.org/web/packages/egg/vignettes/Ecosystem.html)

Examples

if (FALSE) {
data(gwas_CD)
manhattan(gwas_CD)
}