match_alleles {topr}	R Documentation

Match the variants in the snpset by their alleles

Description

match_alleles()

This method is deprecated and will be removed in future versions. use match_by_alleles instead.

Usage

match_alleles(df, verbose = F)

Arguments

df
A dataframe that is in the snpset format (like returned by the get_snpset() function)

verbose
A logical scalar (default: FALSE). Assign to TRUE to get information on which alleles are matched and which are not.

Value

The input dataframe containing only those variants whith matched alleles in the snpset

Examples

## Not run: 
match_alleles(df)

## End(Not run)

[Package topr version 2.0.2 Index]