match_by_alleles {topr}	R Documentation

Match the variants in the snpset by their alleles

Description

match_by_alleles()

Usage

match_by_alleles(df, verbose = NULL, show_full_output = FALSE)

Arguments

df
A dataframe that is in the snpset format (like returned by the get_snpset function)

verbose
A logical scalar (default: FALSE). Assign to TRUE to get information on which alleles are matched and which are not.

show_full_output
A logical scalar (default:FALSE). Assign to TRUE to show the full output from this function

Value

The input dataframe containing only those variants with matched alleles in the snpset

Examples

## Not run: 
CD_UKBB_lead_snps <- get_lead_snps(CD_UKBB)
snpset <- get_snpset(CD_UKBB_lead_snps, CD_FINNGEN)
match_by_alleles(snpset$found)

## End(Not run)

[Package topr version 2.0.2 Index]