match_by_pos {topr}	R Documentation

Get variants that overlap between two datasets

Description

match_by_pos()

Usage

match_by_pos(df1, df2, verbose = NULL, show_full_output = FALSE)

Arguments

df1
A dataframe of variants, has to contain CHROM and POS

df2
A dataframe of variants, has to contain CHROM and POS

verbose
A logical scalar (default: FALSE). Assign to TRUE to get information on which alleles are matched and which are not.

show_full_output
A logical scalar (default:FALSE). Assign to TRUE to show the full output from this function

Value

A list containing two dataframes, one of overlapping snps and the other snps not found in the second input dataset

Examples

## Not run: 
CD_UKBB_index_snps <- get_lead_snps(CD_UKBB)
match_by_pos(CD_UKBB_index_snps, CD_FINNGEN)

## End(Not run)

[Package topr version 2.0.2 Index]