topr {topr}	R Documentation

topr

Description

A package for viewing and annotating genetic association data

topr functions

The main plotting functions are:

manhattan to create Manhattan plot of association results

regionplot to create regional plots of association results for smaller genetic regions

Author(s)

Maintainer: Thorhildur Juliusdottir totajuliusd@gmail.com [copyright holder]

Authors:

Andri Stefansson andrimarstefansson@gmail.com

See Also

Useful links:

https://github.com/totajuliusd/topr

Report bugs at https://github.com/totajuliusd/topr/issues

Examples

library(topr)
# Create a manhattan plot using
manhattan(CD_UKBB)

# Create a regional plot
regionplot(CD_UKBB, gene="IL23R")

[Package topr version 2.0.2 Index]