Finngen v5 Crohn‘s disease (CHRONSMALL)

Dataset retrieved from the Finngen database (version 5) including 968 crohn´s cases and 210,300 controls. The dataset has been filtered on variants with P <1e-03.

CD_FINNGEN

Format

A data frame with 29,926rows and 9 variables:

CHROM

Chromosome, written as for example chr1 or 1

POS

genetic position of the variant

ID

Variant identifier, e.g. rsid

REF

Reference allele

ALT

Alternative allele

SE

Standard Error

P

P-value from Plink run, additive model, reggresion model GLM_FIRTH

beta

Variant effect

cohort_AF

variant allele frequency in the UKBB cohort)

Source

CHRONSMALL_968_cases_210300_controls_p_1e-03