Create a dataframe that can be used as input for making effect plots

create_snpset()

This method is deprecated and will be removed in future versions. use get_snpset instead.

Usage

create_snpset(
  df1,
  df2,
  thresh = 1e-08,
  protein_coding_only = TRUE,
  region_size = 1e+06,
  verbose = F
)

Arguments

df1

The dataframe to extract the top snps from (with p-value below thresh)

df2

The dataframe in which to search for overlapping SNPs from dataframe1

thresh

Numeric, the p-value threshold used for extracting the top snps from dataset 1

protein_coding_only

Logical, set this variable to TRUE to only use protein_coding genes for the annotation

region_size

Integer, the size of the interval which to extract the top snps from

verbose

Logical, (default: FALSE). Assign to TRUE to get information on which alleles are matched and which are not.

Value

Dataframe containing the top hit

Examples

if (FALSE) { # \dontrun{
create_snpset(CD_UKBB,CD_FINNGEN, thresh=1e-09)
} # }