Create a plot comparing effects within two datasets

effect_plot()

This method is deprecated and will be removed in future versions. use effectplot instead.

effect_plot(
  dat,
  pheno_x = "pheno_x",
  pheno_y = "pheno_",
  annotate_with = "Gene_Symbol",
  thresh = 1e-08,
  ci_thresh = 1,
  gene_label_thresh = 1e-08,
  color = get_topr_colors()[1],
  scale = 1
)

Arguments

dat

The input dataframe (snpset) containing one row per variant and P values (P1 and P2) and effects (E1 and E2) from two datasets/phenotypes

pheno_x

A string representing the name of the phenotype whose effect is plotted on the x axis

pheno_y

A string representing the name of the phenotype whose effect is plotted on the y axis

annotate_with

A string, The name of the column that contains the label for the datapoints (default value is Gene_Symbol)

thresh

A number. Threshold cutoff, datapoints with P2 below this threshold are shown as filled circles whereas datapoints with P2 above this threshold are shown as open circles

ci_thresh

A number.Show the confidence intervals if the P-value is below this threshold

gene_label_thresh

A string, label datapoints with P2 below this threshold

color

A string, default value is the first of the topr colors

scale

A number, to change the size of the title and axes labels and ticks at the same time (default = 1)

Examples

if (FALSE) {
effect_plot(dat)
}