Flip to the positive allele for dataset 1

flip_to_positive_allele_for_dat1()

Usage

flip_to_positive_allele_for_dat1(df)

Arguments

df

A dataframe that is in the snpset format (like returned by the get_snpset() function)

Value

The input dataframe after flipping to the positive effect allele in dataframe 1

Examples

if (FALSE) { # \dontrun{
CD_UKBB_index_snps <- get_lead_snps(CD_UKBB)
snpset <- get_snpset(CD_UKBB_index_snps, CD_FINNGEN)
flip_to_positive_allele_for_dat1(snpset$matched)
} # }