Get the genetic position of a gene or genes by their gene name

get_genes_by_Gene_Symbol() Get genes by their gene symbol/name Required parameters is on gene name or a vector of gene names

get_genes_by_Gene_Symbol(genes, chr = NULL)

Arguments

genes
A gene name (e.g. "FTO") or a vector of gene names ( c("FTO","NOD2"))

chr
Search for the genes on this chromsome only

Value

Dataframe of genes

Examples

if (FALSE) {
get_genes_by_Gene_Symbol(c("FTO","THADA"))
}