Get SNPs/variants within region

get_genes_in_region()

get_genes_in_region(
  chr = chr,
  xmin = xmin,
  xmax = xmax,
  protein_coding_only = F,
  show_exons = F,
  show_genes = T,
  build = 38,
  region = NULL
)

Arguments

chr

A string, chromosome (e.g. chr16)

xmin

An integer representing genetic position

xmax

An integer representing genetic position

protein_coding_only

A logical scalar, if TRUE, only protein coding genes are used for annotation

show_exons

Deprecated : A logical scalar, show exons instead of genees (default show_exons=FALSE)

show_genes

A logical scalar, show genes instead of exons (default show_genes=FALSE)

build

A number representing the genome build. Set to 37 to change to build (GRCh37). The default is build 38 (GRCh38).

region

A string representing the genetic region (e.g chr16:50693587-50734041)

Value

the genes the requested region

Examples

get_genes_in_region(region="chr16:50593587-50834041")
#>    chrom gene_start gene_end gene_symbol        biotype
#> 1  chr16      46406    53608      POLR3K protein_coding
#> 2  chr16     425648   523011   RAB11FIP3 protein_coding
#> 3  chr16    5033701  5087379        ALG1 protein_coding
#> 4  chr16    5044121  5066110    C16orf89 protein_coding
#> 5  chr16   50548395 50649249        NKD1 protein_coding
#> 6  chr16   50666299 50681353       SNX20 protein_coding
#> 7  chr16   50693587 50734041        NOD2 protein_coding
#> 8  chr16   50727416 50742815    CYLD-AS1         lncRNA
#> 9  chr16   50742049 50801935        CYLD protein_coding
#> 10 chr16   50742304 50742377     MIR3181          miRNA
#> 11 chr16   50806667 50807629   LINC02168         lncRNA