Create a quantile quantile (QQ) plot

qqtopr() displays QQ plots for association data. Required parameter is at least one dataset (dataframe) containing the association data (with columns CHROM,POS,P

Usage

qqtopr(
  dat,
  scale = 1,
  n_variants = 0,
  breaks = 15,
  title = NULL,
  color = get_topr_colors(),
  size = 1,
  legend_name = "",
  legend_position = "right",
  legend_labels = NULL,
  axis_text_size = 11,
  axis_title_size = 12,
  title_text_size = 13,
  legend_title_size = 12,
  legend_text_size = 12,
  verbose = NULL,
  diagonal_line_color = "#808080"
)

Arguments

dat

Dataframe or a list of dataframes (required columns is P)) of association results.

scale

An integer, plot elements scale, default: 1

n_variants

An integer, total number of variants used in the study

breaks

A number setting the breaks for the axes

title

A string to set the plot title

color

A string or vector of strings setting the color's for the input datasets

size

A number or a vector of numbers, setting the size of the plot points (default: size=1.2)

legend_name

A string, use to change the name of the legend (default: None)

legend_position

A string, top,bottom,left or right

legend_labels

A string or vector of strings representing legend labels for the input datasets

axis_text_size

A number, size of the x and y axes tick labels (default: 12)

axis_title_size

A number, size of the x and y title labels (default: 12)

title_text_size

A number, size of the plot title (default: 13)

legend_title_size

A number, size of the legend title

legend_text_size

A number, size of the legend text

verbose

A logical scalar (default: NULL). Set to FALSE to suppress printed messages

diagonal_line_color

A string setting the color of the diagonal line on the plot

Value

ggplot

Examples

if (FALSE) { # \dontrun{
qqtopr(CD_UKBB)
} # }