Create a QQ plot

qqtopr() displays QQ plots for association data. Required parameter is at least one dataset (dataframe) containing the association data (with columns CHROM,POS,P

qqtopr(
  dat,
  scale = 1,
  n_variants = 0,
  breaks = 15,
  title = NULL,
  color = get_topr_colors(),
  size = 1,
  legend_name = "",
  legend_position = "right",
  legend_labels = NULL,
  axis_text_size = 11,
  axis_title_size = 12,
  title_text_size = 13,
  legend_title_size = 12,
  legend_text_size = 12
)

Arguments

dat

Dataframe or a list of dataframes (required columns is P)) of association results.

scale

An integer, plot elements scale, default: 1

n_variants

An integer, total number of variants used in the study

breaks

A number setting the breaks for the axes

title

A string

color

A string or vector of strings setting the color/s for the input datset/s

size

An integer setting the size of the plot points (default: size=1.2)

legend_name

A string, use to change the name of the legend (default: None)

legend_position

A string, top,bottom,left or right

legend_labels

A string or vector of strings representing legend labels for the input dataset/s

axis_text_size

A number, size of the x and y axes tick labels (default: 12)

axis_title_size

A number, size of the x and y title labels (default: 12)

title_text_size

A number, size of the plot title (default: 13)

legend_title_size

A number, size of the legend title

legend_text_size

A number, size of the legend text

Value

ggplot