Region plot

regionplot() displays the association results for a smaler region within one chromosome Required parameter is at least one dataset (dataframe) containing the association data (with columns CHROM,POS,P in upper or lowercase), xmin and xmax

All other input parameters are optional

regionplot(
  dat,
  annotation_thresh = NULL,
  title = "",
  label_all = 0,
  xmin = 0,
  size = 2,
  shape = 19,
  alpha = 1,
  label_size = 4,
  color = c("darkblue", "#E69F00", "#00AFBB", "#999999", "#FC4E07", "darkorange1"),
  axis_text_size = 11,
  axis_title_size = 12,
  title_text_size = 13,
  legend_title_size = 12,
  legend_text_size = 12,
  show_genes = FALSE,
  show_overview = TRUE,
  show_exons = FALSE,
  sign_thresh = 5.1e-09,
  sign_thresh_color = "red",
  genes = NULL,
  variants = NULL,
  variant_ids = NULL,
  variant_ids_color = "red",
  xmax = NULL,
  ymin = NULL,
  ymax = NULL,
  protein_coding_only = FALSE,
  chr = NULL,
  vline = NULL,
  legend_name = "Data:",
  legend_position = "right",
  legend_labels = NULL,
  gene = NULL,
  highlight_genes = NULL,
  highlight_genes_color = NULL
)

Arguments

annotation_thresh
Display annotation for variants with p-values below this threshold

title
Plot title (optional)

xmin
Parameters setting the chromosomal range to display on the x-axis

size
Optional parameter setting the size of the plot points (default: size=1.2)

shape
A number of vector of numers setting the shape of the plotted points

alpha
A number or vector of numbers setting the transparancy of the plotted points

label_size
Optional parameter to set the size of the plot labels (default: label_size=3.5)

color
Optional parameter setting the color of the plot points (default: color="darkblue")

axis_text_size
Text size of the x and y axes tick labels (default: 12)

axis_title_size
Text size of the x and y title labels (default: 12)

title_text_size
Text size of the plot title (default: 13)

legend_title_size
Text size of the legend title

legend_text_size
Text size of the legend text

show_genes
Show genes instead of exons (default show_genes=FALSE)

show_overview
Show the overview plot (default show_overview=TRUE)

show_exons
Show exons instead of genees (default show_exons=FALSE)#'

sign_thresh
Optional parameter setting the threshold of the dashed red horizontal line representing the significance threshold (default: sign_thresh=5.1e-9). Multiple thresholds can be provided in a vector, e.g sign_thresh=c(5.1e-9,1.0e-6)). Set this parameter to NULL if you dont want this line to appear at all sign_thresh=NULL

sign_thresh_color
set the color of the significance threshold line or lines

genes
List of genes or exons to plot

variants
Optional parameter, a dataset (required columns are CHROM,POS,P) with a list of variants to display and label on the plot. Variants will be labeled with Gene_Symbol if present in the dataframe, otherwise with ID. If neither Gene_Symbol nor ID are in the dataset, the variant will be labelled with its position (CHROM and POS)

variant_ids
A vector of variant ids to highlight on the plot, e.g. variant_ids=c("rs1234, rs45898")

variant_ids_color
The color of the variants in variants_id (default color is red)

xmax
Parameters setting the chromosomal range to display on the x-axis

ymin
Optional parameters, min and max of the y-axis, (default values: ymin=0, ymax=max(-log10(df$P)))

ymax
Optional parameters, min and max of the y-axis, (default values: ymin=0, ymax=max(-log10(df$P)))

protein_coding_only
Set this parameter to TRUE to only use protein coding genes for annotation

chr
The chromosome to plot (i.e. chr15), only required if the input dataframe contains results from more than one chromosome

vline
Parameter (optional) to add a vertical line to the plot at a specific chromosomal position, e.g vline=204000066. Multiple values can be provided in a vector, e.g vline=c(204000066,100500188)

legend_name
Change the name of the legend (default: None)

legend_position
Top,bottom,left or right

legend_labels
Legend labels

highlight_genes
A vector of genes or genes to highlight the datapoints for on the plot

highlight_genes_color
Colors for the hihglighted genes (default: green)

Value

ggplot object

Examples

if (FALSE) {
regionplot(df,xmin=xmin,xmax=xmax)
}